+++ android_activity_lifecycle/CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 0.0.1
 
 Initial release.
+
+## 1.0.0
+
+- Expose activity state as stream
+- Migration: use stateStream field of AndroidActivityLifecycle class instead of setMethodCallHandler